config/requirements.txt
asgiref==3.8.1
cffi==1.17.1
dj-rest-auth==6.0.0
Django==5.1
django-extensions==3.2.3
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.2
sqlparse==0.5.1
